app/requirements.txt
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
alembic==1.12.0
pydantic==2.3.0
pydantic-settings==2.0.3
starlette==0.37.2
fastapi==0.111.0
uvicorn==0.30.1
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
pika==1.3.1